pref: update post image with cloudinary

+++ src/app/post-image/post-image.service.ts
@@ -1,6 +1,11 @@
 import { ERROR_CODE } from '../../interface'
 import { AppError } from '../../middleware'
-import { deleteFile, MESSAGE } from '../../utils'
+import {
+	MESSAGE,
+	uploadImageCloudinary,
+	updateImageCloudinary,
+	deleteImageCloudinary,
+} from '../../utils'
 import * as postRepository from '../post/post.repository'
 
 import type { ICreatePostImage, IUpdatePostImage } from './post-image.interface'
@@ -10,25 +15,27 @@ import {
 } from './post-image.mapper'
 import * as postImageRepository from './post-image.repository'
 
-const PORT = process.env.PORT ?? 3000
-const HOST = process.env.HOST ?? 'http://localhost'
-
 export const createPostImage = async (
 	postId: string,
 	file: Express.Multer.File,
-	fileUrl: string,
 ) => {
-	const { filename, size, path } = file
-	const postImage: ICreatePostImage = {
-		filename,
-		size,
-		path,
-		file_url: `${HOST}:${PORT}/posts/files/${fileUrl}`,
-		post_id: postId,
-	}
 	const post = await postRepository.getPost(postId)
 	if (!post) {
 		return new AppError(ERROR_CODE.NOT_FOUND.code, MESSAGE.ERROR.NOT_FOUND.POST)
+	}
+	const upload = await uploadImageCloudinary(file.path, 'post')
+	if (!upload) {
+		return new AppError(
+			ERROR_CODE.INTERNAL_SERVER_ERROR.code,
+			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.CREATED.POST_IMAGE,
+		)
+	}
+	const postImage: ICreatePostImage = {
+		public_id: upload.public_id,
+		filename: `${upload.public_id}.${upload.format}`,
+		size: upload.bytes,
+		file_url: upload.secure_url,
+		post_id: postId,
 	}
 	const createPostImage = await postImageRepository.createPostImage(postImage)
 	if (!createPostImage) {
@@ -43,21 +50,25 @@ export const createPostImage = async (
 export const updatePostImage = async (
 	postImageId: string,
 	file: Express.Multer.File,
-	fileUrl: string,
 ) => {
-	const { filename, size, path } = file
-	const updatePostImage: IUpdatePostImage = {
-		filename,
-		size,
-		path,
-		file_url: `${HOST}:${PORT}/posts/files/${fileUrl}`,
-	}
 	const postImage = await postImageRepository.getPostImage(postImageId)
 	if (!postImage) {
 		return new AppError(
 			ERROR_CODE.NOT_FOUND.code,
 			MESSAGE.ERROR.NOT_FOUND.POST_IMAGE,
 		)
+	}
+	const upload = await updateImageCloudinary(file.path, postImage.public_id)
+	if (!upload) {
+		return new AppError(
+			ERROR_CODE.INTERNAL_SERVER_ERROR.code,
+			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.UPDATED.POST_IMAGE,
+		)
+	}
+	const updatePostImage: IUpdatePostImage = {
+		filename: `${upload.public_id}.${upload.format}`,
+		size: upload.bytes,
+		file_url: upload.secure_url,
 	}
 	const postImageUpdated = await postImageRepository.updatePostImage(
 		postImageId,
@@ -69,7 +80,6 @@ export const updatePostImage = async (
 			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.UPDATED.POST_IMAGE,
 		)
 	}
-	await deleteFile(postImage.path)
 	return postImageDTOMapper(postImageUpdated)
 }
 
@@ -81,6 +91,13 @@ export const deletePostImage = async (postImageId: string) => {
 			MESSAGE.ERROR.NOT_FOUND.POST_IMAGE,
 		)
 	}
+	const deleteFile = await deleteImageCloudinary(postImage.public_id)
+	if (!deleteFile) {
+		return new AppError(
+			ERROR_CODE.INTERNAL_SERVER_ERROR.code,
+			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.DELETED.POST_IMAGE,
+		)
+	}
 	const postImageDeleted =
 		await postImageRepository.deletePostImage(postImageId)
 	if (!postImageDeleted) {
@@ -89,6 +106,5 @@ export const deletePostImage = async (postImageId: string) => {
 			MESSAGE.ERROR.INTERNAL_SERVER_ERROR.DELETED.POST_IMAGE,
 		)
 	}
-	await deleteFile(postImage.path)
 	return deletePostImageDTOMapper(postImageDeleted)
 }
